Add documentation. Cleanup. Add guardrail not to redirect to current url.
